Moved general functions into better homes
